Update lip-8.md
This is the first version, I added the following:
Prolog that explains about the discovery phase. 
Extended the scope of the fund pull pre-approval.
Added a phrase about revoking authorizations.

Please let me know if it was edited correctly and I will proceed to the "Save my sub account" section. (will require a big change to the doc)